refactor(cartera): exitInvestor usa absorberInversionistaEnCube (helper compartido)
